Continue with adding tests for SocketAddress and tweaking behaviour as needed to behave as I need.

+++ src/address/SocketAddress.h
@@ -56,7 +56,7 @@ namespace kt
 
     std::pair<std::optional<kt::SocketAddress>, int> socketToAddress(const SOCKET&);
 
-    std::pair<std::vector<kt::SocketAddress>, int> resolveToAddresses(const std::optional<std::string>&, const unsigned int&, addrinfo&);
+    std::pair<std::vector<kt::SocketAddress>, int> resolveToAddresses(const std::optional<std::string>&, const unsigned short&, addrinfo&);
 
 #ifdef _WIN32
 	int getAddressLength(const kt::SocketAddress&);
